requirements.txt
Django==5.0.2
psycopg2=2.9.
pillow-10.2.0
ipython
django-dump-load-utf8-0.0.4
pytils-0.4.1